Attempt to reconnect when encountering certain PG errors
The `SidekiqPublisher::ReportUnpublishedCount` will often run
in a long running process meant to capture current stats for `SidekiqPublisher`.

If a connection error occurs (like during a Postgres failover), the gem will not gracefully handle it. The connection gets "poisoned" and
all subsequent calls will continue to use this connection and fail. It will continue to do so until the process gets restarted.

This commit addresses that by reconnecting when encountering this case

Fixes https://github.com/ezcater/sidekiq_publisher/issues/41