transition from defining a Railtie to becoming a Rails Engine
When going through the process of upgrading an application that uses sidekiq_publisher to Ruby 3, we ran into a really weird situation that somehow was traced to the `SidekiqPublisher::Job` AR model.

Essentially, when booting the application, with the `SidekiqPublisher::Job` class included, the application fails to boot, hanging indefinitely in some kind of endless loop inside of native Ruby C code. Tracing things via `lldb` did not provide much more information into what the root cause is, so I went through the process of simply commenting out various gems (starting first with those that have C extensions), eventually removing sidekiq_publisher, when things all of a sudden started working. I then proceeded to do a similar thing with various parts of this gem, before temporarily removing `SidekiqPublisher::Job`, and having that turn out to be the culprit.

What is truly infuriating is that when attempting to do this on a fresh Rails 6.0.3.4 application (the version our application is currently on), this behavior does not occur, which I'd imagine means there is some interplay issues going on with some other gem/code.

I wish I had more specific details about why that class is an issue, but I have yet to reduce the issue down to a minimal reproducible case. What I *was* able to figure out is that removing the explicit `require 'activerecord'` and transitioning this gem to a Rails Engine resolves things, I believe due to the way Rails handles the order of loading models included in an engine.